Command line code to load large large large files into MySQL
Little trick to get a list of all the columns and their corresponding
values. Use the wizard to auto detect the column and values, start the
import and then stop it. It will create an empty table. Then click on
the column info and copy into Excel to add a "," after each variable.